requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# CLI and UI
click>=8.0.0
rich>=13.0.0

# Web frameworks
fastapi>=0.100.0
uvicorn>=0.23.0

# HTTP requests
requests>=2.31.0

# Configuration and data formats
pyyaml>=6.0.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
